Add atomic context with two different lock strategies: original and soft lock (twem-proxy compatible)

+++ cacheops/invalidation.py
@@ -1,6 +1,8 @@
 # -*- coding: utf-8 -*-
+import contextlib
+import time
 import six
-from cacheops.conf import redis_client, handle_connection_failure
+from cacheops.conf import redis_client, handle_connection_failure, USE_SOFT_LOCK, SOFT_LOCK_TIMEOUT
 from cacheops.utils import get_model_name, non_proxy
 
 
@@ -19,6 +21,9 @@ def conj_cache_key(model, conj):
 def conj_cache_key_from_scheme(model, scheme, obj):
     return 'conj:%s:' % get_model_name(model) \
          + '&'.join('%s=%s' % (f, getattr(obj, f)) for f in scheme)
+
+def conj_lock_key(model):
+    return 'conj:%s:lock' % get_model_name(model)
 
 
 class ConjSchemes(object):
@@ -40,15 +45,19 @@ class ConjSchemes(object):
             model_or_name = get_model_name(model_or_name)
         return 'schemes:%s:version' % model_or_name
 
+    def get_schemes_lock_key(self, model_or_name):
+        if not isinstance(model_or_name, six.string_types):
+            model_or_name = get_model_name(model_or_name)
+        return "schemes:%s:lock" % model_or_name
+
     def load_schemes(self, model):
         model_name = get_model_name(model)
 
-        lock_key = redis_lock_acquire("lock:schemes:" + model_name)
-        pipe = redis_client.pipeline(transaction=False)
-        pipe.get(self.get_version_key(model))
-        pipe.smembers(self.get_lookup_key(model_name))
-        version, members = pipe.execute()
-        redis_lock_release(lock_key)
+        lock_key = self.get_schemes_lock_key(model_name)
+        with atomic_redis_pipeline(lock_key) as txn:
+            txn.get(self.get_version_key(model))
+            txn.smembers(self.get_lookup_key(model_name))
+            version, members = txn.execute()
 
         self.local[model_name] = set(map(deserialize_scheme, members))
         self.local[model_name].add(()) # Всегда добавляем пустую схему
@@ -86,23 +95,13 @@ class ConjSchemes(object):
                 schemes = self.load_schemes(model)
             if new_schemes - schemes:
                 # Write new schemes to redis
-                # txn = redis_client.pipeline()
-                # txn.incr(self.get_version_key(model_name)) # Увеличиваем версию схем
-                #
-                # lookup_key = self.get_lookup_key(model_name)
-                # for scheme in new_schemes - schemes:
-                #     txn.sadd(lookup_key, serialize_scheme(scheme))
-                # txn.execute()
-                lock_key = redis_lock_acquire("lock:schemes:" + model_name)
-                pipe = redis_client.pipeline(transaction=False)
-                pipe.incr(self.get_version_key(model_name))
-
-                lookup_key = self.get_lookup_key(model_name)
-                for scheme in new_schemes - schemes:
-                    pipe.sadd(lookup_key, serialize_scheme(scheme))
-                pipe.execute()
-                redis_lock_release(lock_key)
-
+                lock_key = self.get_schemes_lock_key(model_name)
+                with atomic_redis_pipeline(lock_key) as txn:
+                    txn.incr(self.get_version_key(model_name))
+                    lookup_key = self.get_lookup_key(model_name)
+                    for scheme in new_schemes - schemes:
+                        txn.sadd(lookup_key, serialize_scheme(scheme))
+                        txn.execute()
 
                 # Updating local version
                 self.local[model_name].update(new_schemes)
@@ -126,17 +125,35 @@ class ConjSchemes(object):
 cache_schemes = ConjSchemes()
 
 
-def redis_lock_acquire(lock_key, timeout_ms=1000, redis_client=redis_client):
-    """ Spin-lock, be aware!
+class LockWaitTimeout(Exception):
+    """ Error raised if soft lock could not be acquired for limited time.
     """
-    while not redis_client.set(lock_key, "1",  nx=True, px=timeout_ms):
-        pass
-    return lock_key
 
 
-def redis_lock_release(lock_key, redis_client=redis_client):
-    redis_client.delete(lock_key)
-    return
+@contextlib.contextmanager
+def atomic_redis_pipeline(lock_key, timeout_ms=SOFT_LOCK_TIMEOUT,
+                          redis_client=redis_client):
+    """ Context manager which provide atomic Redis pipeline.
+
+    There are two different atomic strategies:
+        - Redis native transaction by default
+        - Setting lock key if CACHEOPS_USE_SOFT_LOCK = True
+
+    May raise LockWaitTimeout if can`t get lock for 2 * timeout_ms
+    """
+    try:
+        if USE_SOFT_LOCK:
+            start_time = time.time()
+            while not redis_client.set(lock_key, "1",  nx=True, px=timeout_ms):
+                # Something like spin-lock, may be highly CPU expensive.
+                if int(time.time() - start_time) * 1000 > 2 * timeout_ms:
+                    # Prevent infinite loop
+                    raise LockWaitTimeout(lock_key)
+        # If soft lock is used then transaction is not needed
+        yield redis_client.pipeline(transaction=not USE_SOFT_LOCK)
+    finally:
+        if USE_SOFT_LOCK:
+            redis_client.delete(lock_key)
 
 @handle_connection_failure
 def invalidate_obj(obj):
@@ -156,38 +173,37 @@ def invalidate_obj(obj):
 
         # Reading scheme version, cache_keys and deleting invalidators in
         # a single transaction.
-        # def _invalidate_conjs(pipe):
-        #     # get schemes version to check later that it's not obsolete
-        #     pipe.get(cache_schemes.get_version_key(model))
-        #     # Get a union of all cache keys registered in invalidators
-        #     pipe.sunion(conjs_keys)
-        #     # `conjs_keys` are keys of sets containing `cache_keys` we are going to delete,
-        #     # so we'll remove them too.
-        #     # NOTE: There could be some other invalidators not matched with current object,
-        #     #       which reference cache keys we delete, they will be hanging out for a while.
-        #     pipe.delete(*conjs_keys)
+        try:
+            lock_key = conj_lock_key(model)
+            with atomic_redis_pipeline(lock_key) as pipe:
+                # get schemes version to check later that it's not obsolete
+                pipe.get(cache_schemes.get_version_key(model))
+                # Get a union of all cache keys registered in invalidators
+                pipe.sunion(conjs_keys)
+                # `conjs_keys` are keys of sets containing `cache_keys` we are going to delete,
+                # so we'll remove them too.
+                # NOTE: There could be some other invalidators not matched with current object,
+                #       which reference cache keys we delete, they will be hanging out for a while.
+                pipe.delete(*conjs_keys)
 
-        for ck in conjs_keys:
-            lock_key = redis_lock_acquire("lock:" + ck)
+                # NOTE: we delete fetched cache_keys later which makes a tiny possibility that something
+                #       will fail in the middle leaving those cache keys hanging without invalidators.
+                #       The alternative WATCH-based optimistic locking proved to be pessimistic.
+                version, cache_keys, _ = pipe.execute()
+        except LockWaitTimeout:
+            # Looks no good, but we still can clear cache_keys. Conjs_keys
+            # will be cleared later, hope so.
             pipe = redis_client.pipeline(transaction=False)
-            pipe.smembers(ck)
-            pipe.delete(ck)
-            cache_keys, _ = pipe.execute()
-            redis_lock_release(lock_key)
-            if cache_keys:
-                redis_client.delete(*cache_keys)
+            pipe.get(cache_schemes.get_version_key(model))
+            pipe.sunion(conjs_keys)
+            version, cache_keys = pipe.execute()
 
-        # NOTE: we delete fetched cache_keys later which makes a tiny possibility that something
-        #       will fail in the middle leaving those cache keys hanging without invalidators.
-        #       The alternative WATCH-based optimistic locking proved to be pessimistic.
-        # version, cache_keys, _ = redis_client.transaction(_invalidate_conjs)
-        # if cache_keys:
-        #     redis_client.delete(*cache_keys)
+        if cache_keys:
+            redis_client.delete(*cache_keys)
 
         # OK, we invalidated all conjunctions we had in memory, but model schema
         # may have been changed in redis long time ago. If this happened,
         # schema version will not match and we should load new schemes and redo our thing
-        version = redis_client.get(cache_schemes.get_version_key(model))
         if int(version or 0) != cache_schemes.version(model):
             schemes = cache_schemes.load_schemes(model)
         else:
